Moves getGeographyById from data/bin to a util file.
