+++ src/components/OUSCard.tsx
@@ -21,7 +21,7 @@ import {
   toPercentMetric,
 } from "../../data/bin/getPrecalcMetrics";
 import { GeoProp } from "../util/types";
-import { getGeographyById } from "../util/geography";
+import { getGeographyById } from "../util/getGeographyById";
 
 export const OUSCard: React.FunctionComponent<GeoProp> = (props) => {
   const [{ isCollection }] = useSketchProperties();
