+++ src/functions/ousDemographicOverlap.ts
@@ -15,7 +15,10 @@ import {
 } from "../util/overlapOusDemographic";
 import { featureCollection } from "@turf/helpers";
 import project from "../../project";
-import { clipSketchToGeography, getParamStringArray } from "../util/geography";
+import {
+  clipSketchToGeography,
+  getParamStringArray,
+} from "../util/clipSketchToGeography";
 import { ExtraParams } from "../util/types";
 
 const METRIC = project.getMetricGroup("ousSectorDemographicOverlap");
